requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.60.0
numpy>=1.21.0
Pillow>=8.0.0
seaborn>=0.11.0
opencv-python>=4.5.0

